pyproject.toml
[tool.poetry]
name = "breastCancerDataMining"
version = "0.1.0"
description = "Data mining assignment to Discover patterns on benign and malignant breast cancer"
authors = ["Nagi0 <brunolefonseca13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
polars = "^1.18.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
tqdm = "^4.67.1"
ucimlrepo = "^0.0.7"
scikit-learn = "^1.5.2"
llvmlite = "^0.43.0"
umap-learn = "^0.5.7"
orange3 = "^3.38.1"
plotly = "^5.24.1"
matplotlib = "^3.10.0"
ipykernel = "^6.29.5"
dtreeviz = "^2.2.2"
graphviz = "^0.20.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
